added a summary of BOLTs PR 918
Of course one can argue whether this is a notable code change. 
I think this feature is very useful for LSPs and it seemed to be heavily motivated from that context by Matt. As Optech reader are mainly professionals I thought I should at least suggest to write about the update.